+++ include/build_info.h
@@ -1,4 +1,4 @@
 #pragma once
 
 constexpr const char* BUILD_ID =
-    "7IN-20260801-PRODUCT54-LOCAL-WEB-OTA";
+    "7IN-20260801-PRODUCT55-AIRSPACE";